changes from https://github.com/broadinstitute/ddp-angular/pull/1259/files
