+++ ddp-workspace/projects/ddp-dsm-ui/src/app/abstraction-field/abstraction-field.component.html
@@ -39,7 +39,7 @@
           'Abstraction-Background-Check': field.qcWrapper.check && (type === 'abstraction' || type === 'review'), 'Cell--Bottom-Border': field.type === 'medication' || field.type === 'recurrence'}">
       <b>{{field.displayName}}</b>
       <div *ngIf="participant[activityOfField] != null && participant[activityOfField].colorNotFinished && (!field.fieldValue.noData
-      && (field.fieldValue.value === ''|| field.fieldValue.value === null || field.fieldValue.value == undefined))"
+      && (field.fieldValue.value === ''|| field.fieldValue.value == null || field.fieldValue.value?.length === 0))"
            class="Color--warn"><b>Field not completed</b></div>
       <a href="#" *ngIf="field.helpText != null && field.helpText !== '' && (type !== 'abstraction' && type !== 'review')"
          [attr.aria-expanded]="field.viewHelpText" [attr.aria-controls]="field.displayName+'_Help'"
