+++ ddp-workspace/projects/ddp-dsm-ui/src/app/abstraction-group/abstraction-group.model.ts
@@ -1,5 +1,5 @@
-import { AbstractionField } from '../medical-record-abstraction/medical-record-abstraction-field.model';
-import {Abstraction} from '../medical-record-abstraction/medical-record-abstraction.model';
+import { AbstractionField } from '../medical-record-abstraction/model/medical-record-abstraction-field.model';
+import {Abstraction} from '../medical-record-abstraction/model/medical-record-abstraction.model';
 
 export class AbstractionGroup {
   constructor(
