+++ ddp-workspace/projects/ddp-dsm-ui/src/app/medical-record-abstraction/medical-record-abstraction-field.model.ts
@@ -1,4 +1,6 @@
-import { Value } from '../utils/value.model';
+import { Value } from '../../utils/value.model';
+import { QCWrapper } from './qc-wrapper.model';
+import { AbstractionFieldValue } from './abstraction-field-value.model';
 
 export class AbstractionField {
   newAdded = false;
@@ -11,18 +13,7 @@ export class AbstractionField {
   constructor(public medicalRecordAbstractionFieldId: number, public displayName: string, public type: string,
               public helpText: string, public orderNumber: number,
               public possibleValues: Value[], public additionalType: string, public hide: boolean,
-              // eslint-disable-next-line max-len
-              public fieldValue: AbstractionFieldValue, public qcWrapper: QCWrapper, public fileInfo: boolean) { // last line are value fields!
-    this.medicalRecordAbstractionFieldId = medicalRecordAbstractionFieldId;
-    this.displayName = displayName;
-    this.helpText = helpText;
-    this.type = type;
-    this.orderNumber = orderNumber;
-    this.possibleValues = possibleValues;
-    this.additionalType = additionalType;
-    this.fieldValue = fieldValue;
-    this.qcWrapper = qcWrapper;
-    this.fileInfo = fileInfo;
+              public fieldValue: AbstractionFieldValue, public qcWrapper: QCWrapper, public fileInfo: boolean) {
   }
 
   static parse(json): AbstractionField {
@@ -68,46 +59,5 @@ export class AbstractionField {
       }
     }
     return cleanedSettings;
-  }
-}
-
-// eslint-disable-next-line max-classes-per-file
-export class AbstractionFieldValue {
-  constructor(public medicalRecordAbstractionFieldId: number, public primaryKeyId: number, public value: string | string[],
-              public valueCounter: number, public note: string, public question: string, public noData: boolean,
-              public doubleCheck: boolean, public filePage: string, public fileName: string, public matchPhrase: string) {
-    this.medicalRecordAbstractionFieldId = medicalRecordAbstractionFieldId;
-    this.primaryKeyId = primaryKeyId;
-    this.value = value;
-    this.valueCounter = valueCounter;
-    this.note = note;
-    this.question = question;
-    this.noData = noData;
-    this.doubleCheck = doubleCheck;
-    this.filePage = filePage;
-    this.fileName = fileName;
-    this.matchPhrase = matchPhrase;
-  }
-
-  static parse(json): AbstractionFieldValue {
-    return new AbstractionFieldValue(
-      json.medicalRecordAbstractionFieldId, json.primaryKeyId, json.value, json.valueCounter, json.note, json.question,
-      json.noData, json.doubleCheck, json.filePage, json.fileName, json.matchPhrase
-    );
-  }
-}
-
-// eslint-disable-next-line max-classes-per-file
-export class QCWrapper {
-  constructor(public abstraction: AbstractionFieldValue, public review: AbstractionFieldValue,
-              public equals: boolean, public check: boolean) {
-    this.abstraction = abstraction;
-    this.review = review;
-    this.equals = equals;
-    this.check = check;
-  }
-
-  static parse(json): QCWrapper {
-    return new QCWrapper(json.abstraction, json.review, json.equals, json.check);
   }
 }
