+++ ddp-workspace/projects/ddp-dsm-ui/src/app/medical-record-abstraction/medical-record-abstraction.model.ts
@@ -3,14 +3,6 @@ export class Abstraction {
 
   constructor(public medicalRecordAbstractionActivityId: number, public participantId: string, public user: string, public activity: string,
               public aStatus: string, public startDate: number, public lastChanged: number, public filesUsed: string) {
-    this.medicalRecordAbstractionActivityId = medicalRecordAbstractionActivityId;
-    this.participantId = participantId;
-    this.user = user;
-    this.activity = activity;
-    this.aStatus = aStatus;
-    this.startDate = startDate;
-    this.lastChanged = lastChanged;
-    this.filesUsed = filesUsed;
   }
 
   static parse(json): Abstraction {
