+++ ddp-workspace/projects/ddp-dsm-ui/src/app/medical-record-abstraction/model/qc-wrapper.model.ts
@@ -1,0 +1,11 @@
+import { AbstractionFieldValue } from './abstraction-field-value.model';
+
+export class QCWrapper {
+  constructor(public abstraction: AbstractionFieldValue, public review: AbstractionFieldValue,
+              public equals: boolean, public check: boolean) {
+  }
+
+  static parse(json): QCWrapper {
+    return new QCWrapper(json.abstraction, json.review, json.equals, json.check);
+  }
+}
