+++ ddp-workspace/projects/ddp-dsm-ui/src/app/utils/utils.ts
@@ -12,7 +12,7 @@ import { QuestionAnswer } from '../activity-data/models/question-answer.model';
 import { QuestionDefinition } from '../activity-data/models/question-definition.model';
 import { FieldSettings } from '../field-settings/field-settings.model';
 import { Filter } from '../filter-column/filter-column.model';
-import { AbstractionField } from '../medical-record-abstraction/medical-record-abstraction-field.model';
+import { AbstractionField } from '../medical-record-abstraction/model/medical-record-abstraction-field.model';
 import { Participant } from '../participant-list/participant-list.model';
 import { NameValue } from './name-value.model';
 import { AnswerValue } from '../../../../ddp-sdk/src/lib/models/activity/answerValue';
